:recycle: Detach getUserProfile as independent service
to prevent circular reference
