+++ jogakbo/src/main/java/com/noyes/jogakbo/user/UserService.java
@@ -37,7 +37,6 @@ public class UserService {
   private final UserRepository userRepository;
   private final JwtService jwtService;
   private final AwsS3Service awsS3Service;
-  private final AlbumService albumService;
 
   public Optional<User> checkUser(HttpServletResponse response, String accessToken) {
 
@@ -165,48 +164,6 @@ public class UserService {
         .userUUID(user.getUserUUID())
         .nickname(user.getNickname())
         .profileImageURL(user.getProfileImageUrl()).build();
-  }
-
-  /**
-   * Id에 해당하는 User의 Profile 반환
-   * 클래스 내의 getUser() Method를 사용하여 특정 User 객체를 얻어 처리함
-   * 
-   * @param userUUID
-   * @return
-   */
-  public UserProfile getUserProfile(String userUUID) {
-
-    User user = getUser(userUUID);
-
-    // 친구들 정보 불러오기
-    List<String> friendsUUIDs = user.getFriends();
-    List<Friend> friends = getFriends(friendsUUIDs);
-
-    // 친구 요청을 보낸 유저 정보 불러오기
-    List<String> friendRequestersUUIDs = user.getFriendRequesters();
-    List<Friend> friendRequesters = getFriends(friendRequestersUUIDs);
-
-    // 소유하고 있는 앨범 정보 불러오기
-    List<String> albumsUUIDs = user.getAlbums();
-    List<AlbumInfo> albums = albumService.getAlbumInfo(albumsUUIDs);
-
-    // 공동 작업 중인 앨범 정보 불러오기
-    List<String> collaboAlbumsUUIDs = user.getCollaboAlbums();
-    List<AlbumInfo> collaboAlbums = albumService.getAlbumInfo(collaboAlbumsUUIDs);
-
-    // 초대받은 앨범 정보 불러오기
-    List<String> albumInvitersUUIDs = user.getAlbumInviters();
-    List<AlbumInfo> albumInviters = albumService.getAlbumInfo(albumInvitersUUIDs);
-
-    return UserProfile.builder()
-        .nickname(user.getNickname())
-        .profileImageUrl(user.getProfileImageUrl())
-        .friends(friends)
-        .receivedFriendRequest(friendRequesters)
-        .albums(albums)
-        .collaboAlbums(collaboAlbums)
-        .receivedAlbumInvitations(albumInviters)
-        .build();
   }
 
   /**
